Conflicts:
	README.md
	trunk/user/accel-pptpd/pptpd-1.3.3/if_pppox.h
	trunk/user/minidlna/minidlna-1.x.x/configure
	trunk/versions.inc

+++ trunk/user/httpd/web_ex.c
@@ -56,13 +56,13 @@
 #include "nvram_x.h"
 #include "httpd.h"
 
-#define MAX_GROUP_COUNT		64
 #define GROUP_FLAG_REFRESH 	0
 #define GROUP_FLAG_DELETE 	1
 #define GROUP_FLAG_ADD 		2
 #define GROUP_FLAG_REMOVE 	3
 
 static int apply_cgi_group(webs_t wp, int sid, struct variable *var, const char *groupName, int flag);
+static void nvram_clr_group_temp(struct variable *v);
 static int nvram_generate_table(webs_t wp, char *serviceId, char *groupName);
 
 static int nvram_modified = 0;
@@ -75,7 +75,7 @@ static u64 restart_needed_bits = 0;
 static char post_buf[32768] = {0};
 static char next_host[128] = {0};
 static char SystemCmd[128] = {0};
-static int delMap[MAX_GROUP_COUNT+2];
+static int  group_del_map[MAX_GROUP_COUNT+2];
 
 extern struct evDesc events_desc[];
 extern int auth_nvram_changed;
@@ -394,10 +394,12 @@ websScan(const char *query)
 		
 		if (v2 != NULL && ((sp = strchr(v1+1, ' ')) == NULL || (sp > v2))) {
 			if (flag && strncmp(v1+1, groupid, strlen(groupid))==0) {
-				delMap[i] = atoi(value);
-				if (delMap[i]==-1)
-					break;
-				i++;
+				if (i < MAX_GROUP_COUNT) {
+					group_del_map[i] = atoi(value);
+					if (group_del_map[i]==-1)
+						break;
+					i++;
+				}
 			} else if (strncmp(v1+1, "group_id", 8)==0) {
 				snprintf(groupid, sizeof(groupid), "%s_s", value);
 				flag = 1;
@@ -406,7 +408,7 @@ websScan(const char *query)
 		v1 = strchr(v1+1, '&');
 	}
 
-	delMap[i] = -1;
+	group_del_map[i] = -1;
 }
 
 static void
@@ -767,28 +769,6 @@ svc_pop_list(char *value, char key)
 #define WIFI_GUEST_CONTROL_BIT	(1<<3)
 #define WIFI_SCHED_CONTROL_BIT	(1<<4)
 
-static const char* wifn_list[][3] = {
-	{IFNAME_2G_MAIN, IFNAME_2G_APCLI, IFNAME_2G_WDS0},
-#if BOARD_HAS_5G_RADIO
-	{IFNAME_5G_MAIN, IFNAME_5G_APCLI, IFNAME_5G_WDS0}
-#endif
-};
-
-static char*
-get_wifi_ifname(int is_5g)
-{
-	int i;
-
-	is_5g &= 1;
-	for (i = 0; i < ARRAY_SIZE(wifn_list[is_5g]); i++) {
-		char *wifn = (char *)wifn_list[is_5g][i];
-		if (is_interface_up(wifn))
-			return wifn;
-	}
-
-	return NULL;
-}
-
 static void
 set_wifi_param_int(const char* ifname, char* param, char* value, int val_min, int val_max)
 {
@@ -978,7 +958,7 @@ validate_asp_apply(webs_t wp, int sid)
 			
 			if (!strcmp(v->name, "wl_TxPower"))
 			{
-				char *wifn = get_wifi_ifname(1);
+				const char *wifn = find_wlan_if_up(1);
 				if (wifn)
 					set_wifi_param_int(wifn, "TxPower", value, 0, 100);
 				
@@ -1041,7 +1021,7 @@ validate_asp_apply(webs_t wp, int sid)
 			
 			if (!strcmp(v->name, "rt_TxPower"))
 			{
-				char *wifn = get_wifi_ifname(0);
+				const char *wifn = find_wlan_if_up(0);
 				if (wifn)
 					set_wifi_param_int(wifn, "TxPower", value, 0, 100);
 				
@@ -1053,6 +1033,16 @@ validate_asp_apply(webs_t wp, int sid)
 				
 				rt_modified |= WIFI_IWPRIV_CHANGE_BIT;
 			}
+#if (BOARD_NUM_UPHY_USB3 > 0)
+			else if (!strcmp(v->name, "rt_VgaClamp"))
+			{
+				const char *wifn = find_wlan_if_up(0);
+				if (wifn)
+					set_wifi_param_int(wifn, "VgaClamp", value, 0, 4);
+				
+				rt_modified |= WIFI_IWPRIV_CHANGE_BIT;
+			}
+#endif
 #if defined(USE_RT3352_MII)
 			else if (!strcmp(v->name, "rt_IgmpSnEnable"))
 			{
@@ -1215,7 +1205,7 @@ update_variables_ex(int eid, webs_t wp, int argc, char **argv)
 					
 					validate_asp_apply(wp, sid);	// for some nvram with this group
 					
-					if (nvram_get_int(group_id) > 0) {
+					if (v->name && nvram_get_int(group_id) > 0) {
 						restart_needed_bits |= (v->event_mask & ~(EVM_BLOCK_UNSAFE));
 						dbG("group restart_needed_bits: 0x%llx\n", restart_needed_bits);
 #if BOARD_HAS_5G_RADIO
@@ -1227,6 +1217,7 @@ update_variables_ex(int eid, webs_t wp, int argc, char **argv)
 						
 						nvram_modified = 1;
 						nvram_set_int_temp(group_id, 0);
+						nvram_clr_group_temp(v);
 					}
 					
 					if (nvram_modified)
@@ -2040,6 +2031,9 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 #else
 	int has_ipv6 = 0;
 #endif
+
+#if defined(USE_HW_NAT)
+	int has_ipv4_ppe = 1;
 #if defined(USE_IPV6_HW_NAT)
 #if defined(USE_HW_NAT_V2)
 	int has_ipv6_ppe = 2;
@@ -2049,6 +2043,10 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 #else
 	int has_ipv6_ppe = 0;
 #endif
+#else
+	int has_ipv4_ppe = 0;
+#endif
+
 #if (BOARD_RAM_SIZE < 64)
 	int max_conn = 16384;
 #elif (BOARD_RAM_SIZE < 128)
@@ -2056,7 +2054,7 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 #elif (BOARD_RAM_SIZE < 256)
 	int max_conn = 262144;
 #else
-	int max_conn = 524288;
+	int max_conn = 327680;
 #endif
 #if defined (USE_NAND_FLASH)
 	int has_mtd_rwfs = 1;
@@ -2155,6 +2153,7 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 	websWrite(wp,
 		"function support_ipv6() { return %d;}\n"
 		"function support_ipv6_ppe() { return %d;}\n"
+		"function support_ipv4_ppe() { return %d;}\n"
 		"function support_peap_ssl() { return %d;}\n"
 		"function support_http_ssl() { return %d;}\n"
 		"function support_ddns_ssl() { return %d;}\n"
@@ -2164,6 +2163,7 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 		"function support_btn_mode() { return %d;}\n"
 		"function support_usb() { return %d;}\n"
 		"function support_usb3() { return %d;}\n"
+		"function support_num_usb() { return %d;}\n"
 		"function support_switch_type() { return %d;}\n"
 		"function support_num_ephy() { return %d;}\n"
 		"function support_ephy_w1000() { return %d;}\n"
@@ -2178,6 +2178,7 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 		"function support_2g_stream_rx() { return %d;}\n",
 		has_ipv6,
 		has_ipv6_ppe,
+		has_ipv4_ppe,
 		has_peap_ssl,
 		has_http_ssl,
 		has_ddns_ssl,
@@ -2187,6 +2188,7 @@ ej_firmware_caps_hook(int eid, webs_t wp, int argc, char **argv)
 		has_btn_mode,
 		has_usb,
 		has_usb3,
+		BOARD_NUM_USB_PORTS,
 		has_switch_type,
 		BOARD_NUM_ETH_EPHY,
 		BOARD_HAS_EPHY_W1000,
@@ -2212,10 +2214,15 @@ ej_hardware_pins_hook(int eid, webs_t wp, int argc, char **argv)
 #else
 	int has_but_wps = 0;
 #endif
-#if defined (BOARD_GPIO_BTN_WLTOG)
-	int has_but_wlt = 1;
+#if defined (BOARD_GPIO_BTN_FN1)
+	int has_but_fn1 = 1;
 #else
-	int has_but_wlt = 0;
+	int has_but_fn1 = 0;
+#endif
+#if defined (BOARD_GPIO_BTN_FN2)
+	int has_but_fn2 = 1;
+#else
+	int has_but_fn2 = 0;
 #endif
 #if defined (BOARD_GPIO_LED_ALL)
 	int has_led_all = 1;
@@ -2250,7 +2257,8 @@ ej_hardware_pins_hook(int eid, webs_t wp, int argc, char **argv)
 
 	websWrite(wp,
 		"function support_but_wps() { return %d;}\n"
-		"function support_but_wlt() { return %d;}\n"
+		"function support_but_fn1() { return %d;}\n"
+		"function support_but_fn2() { return %d;}\n"
 		"function support_led_all() { return %d;}\n"
 		"function support_led_wan() { return %d;}\n"
 		"function support_led_lan() { return %d;}\n"
@@ -2259,7 +2267,8 @@ ej_hardware_pins_hook(int eid, webs_t wp, int argc, char **argv)
 		"function support_led_pwr() { return %d;}\n"
 		"function support_led_phy() { return %d;}\n",
 		has_but_wps,
-		has_but_wlt,
+		has_but_fn1,
+		has_but_fn2,
 		has_led_all,
 		has_led_wan,
 		has_led_lan,
@@ -2978,71 +2987,83 @@ apply_cgi(const char *url, webs_t wp)
 	return 1;
 }
 
-void nvram_add_group_item(webs_t wp, struct variable *v, int sid)
+static void
+nvram_clr_group_temp(struct variable *v)
 {
-    struct variable *gv;
-    char *value;
-    char name[64];
-    int count;
+	char name[64];
+	struct variable *gv;
 
-    if (v->argv[0]==NULL) 
-    {
-       return;
-    }
+	if (v->argv[0]==NULL)
+		return;
 
-    count = nvram_get_int(v->argv[3]);
-
-    for (gv = (struct variable *)v->argv[0]; gv->name!=NULL; gv++)
-    {
-	snprintf(name, sizeof(name), "%s_0", gv->name);
-	if ((value=websGetVar(wp, name, NULL)))
-	{
-	    nvram_add_list_x(gv->name, value, count);
+	for (gv = (struct variable *)v->argv[0]; gv->name!=NULL; gv++) {
+		snprintf(name, sizeof(name), "%s_0", gv->name);
+		
+		/* clear last deleted value */
+		nvram_unset(name);
 	}
-	else
-	{
-	    nvram_add_list_x(gv->name, "", count);
-	}
-    }
-
-    count++;
-    nvram_set_int(v->argv[3], count);
 }
 
-
-void nvram_remove_group_item(webs_t wp, struct variable *v, int sid, int *delMap)
+static void
+nvram_add_group_item(webs_t wp, struct variable *v)
 {
-    struct variable *gv;
-    int i, deleted, gcount;
+	char name[64], *value;
+	struct variable *gv;
+	int gcount;
 
-    if (v->argv[0]==NULL) 
-    {
-       return;
-    }
+	if (v->argv[0]==NULL)
+		return;
 
-    gcount = nvram_get_int(v->argv[3]);
-    if (gcount < 1)
-    {
-       return;
-    }
+	gcount = nvram_get_int(v->argv[3]);
+	if (gcount < 0)
+		gcount = 0;
 
-    for (gv = (struct variable *)v->argv[0]; gv->name!=NULL; gv++)
-    {
-        nvram_del_list_map_x(gv->name, gcount, delMap);
-    }
+	for (gv = (struct variable *)v->argv[0]; gv->name!=NULL; gv++) {
+		snprintf(name, sizeof(name), "%s_0", gv->name);
+		
+		/* clear last deleted value */
+		nvram_unset(name);
+		
+		value = websGetVar(wp, name, "");
+		snprintf(name, sizeof(name), "%s%d", gv->name, gcount);
+		nvram_set(name, value);
+	}
 
-    deleted = 0;
-    for (i=0; i < gcount; i++)
-    {
-        if (delMap[deleted]==i)
-            deleted++;
-    }
+	gcount++;
+	nvram_set_int(v->argv[3], gcount);
+}
 
-    if (deleted > 0)
-    {
-        gcount -= deleted;
-        nvram_set_int(v->argv[3], gcount);
-    }
+static void
+nvram_remove_group_item(struct variable *v, int *delMap)
+{
+	struct variable *gv;
+	int i, di, gcount;
+
+	if (v->argv[0]==NULL)
+		return;
+
+	gcount = nvram_get_int(v->argv[3]);
+	if (gcount < 1)
+		return;
+
+	for (gv = (struct variable *)v->argv[0]; gv->name!=NULL; gv++)
+		nvram_del_list_map_x(gv->name, gcount, delMap);
+
+	di = 0;
+	for (i=0; i < gcount; i++) {
+		if (delMap[di]==i) {
+			di++;
+			if (di > MAX_GROUP_COUNT)
+				break;
+		}
+	}
+
+	if (di > 0) {
+		gcount -= di;
+		if (gcount < 0)
+			gcount = 0;
+		nvram_set_int(v->argv[3], gcount);
+	}
 }
 
 /* Rule for table: 
@@ -3146,73 +3167,67 @@ ToHTML:
 static int
 apply_cgi_group(webs_t wp, int sid, struct variable *var, const char *groupName, int flag)
 {
-   struct variable *v;
-   int groupCount;
+	struct variable *v;
 
-   if (var!=NULL)
-     v = var;
-   else
-   {
-       /* Validate and set vairables in table order */
-       for (v = GetVariables(sid); v->name != NULL; v++)
-       {
-       	   if (!strcmp(groupName, v->name)) 
-       		break;
-       }
-   }
+	if (var != NULL) {
+		v = var;
+	} else {
+		/* Validate and set vairables in table order */
+		for (v = GetVariables(sid); v->name != NULL; v++) {
+			if (!strcmp(groupName, v->name))
+				break;
+		}
+	}
 
-   if (v->name == NULL) return 0;
+	if (v->name == NULL)
+		return 0;
 
-   groupCount = atoi(v->argv[1]);
+	if (flag == GROUP_FLAG_ADD) {
+		nvram_add_group_item(wp, v);
+		return 1;
+	} else if (flag == GROUP_FLAG_REMOVE) {
+		nvram_remove_group_item(v, group_del_map);
+		return 1;
+	}
 
-   if (flag == GROUP_FLAG_ADD)/* if (!strcmp(value, " Refresh ")) || Save condition */
-   {
-	nvram_add_group_item(wp, v, sid);
-	return 1;	// 2008.08 magic
-   }
-   else if (flag == GROUP_FLAG_REMOVE)/* if (!strcmp(value, " Refresh ")) || Save condition */
-   {
-   	nvram_remove_group_item(wp, v, sid, delMap);  
-	return 1; 	// 2008.08 magic
-   }
-  	return 0; // 2008.08 magic
+	return 0;
 }
 
 static int
 nvram_generate_table(webs_t wp, char *serviceId, char *groupName)
 {
-   struct variable *v;
-   int i, groupCount, ret, r, sid;
+	struct variable *v;
+	int i, groupCount, ret, r, sid;
 
-   sid = LookupServiceId(serviceId);
+	sid = LookupServiceId(serviceId);
+	if (sid == -1)
+		return 0;
 
-   if (sid==-1) return 0;
+	/* Validate and set vairables in table order */
+	for (v = GetVariables(sid); v->name != NULL; v++) {
+		if (!strcmp(groupName, v->name))
+			break;
+	}
 
-   /* Validate and set vairables in table order */
-   for (v = GetVariables(sid); v->name != NULL; v++) 
-   {
-      if (!strcmp(groupName, v->name)) break;
-   }
+	if (v->name == NULL)
+		return 0;
 
-   if (v->name == NULL) return 0;    
+	groupCount = nvram_get_int(v->argv[3]);
 
-   groupCount = nvram_get_int(v->argv[3]);
+	if (groupCount == 0) {
+		ret = nvram_add_group_table(wp, serviceId, v, -1);
+	} else {
+		ret = 0;
+		for (i=0; i<groupCount; i++) {
+			r = nvram_add_group_table(wp, serviceId, v, i);
+			if (r != 0)
+				ret += r;
+			else
+				break;
+		}
+	}
 
-   if (groupCount==0)
-       ret = nvram_add_group_table(wp, serviceId, v, -1);
-   else
-   {
-      ret = 0;
-      for (i=0; i<groupCount; i++)
-      {
-	r = nvram_add_group_table(wp, serviceId, v, i);
-	if (r!=0)
-	   ret += r;
-	else break;
-      }
-   }
-
-   return (ret);
+	return (ret);
 }
 
 void
@@ -3242,6 +3257,8 @@ static void
 do_html_apply_post(const char *url, FILE *stream, int clen, char *boundary)
 {
 	init_cgi(NULL);
+
+	group_del_map[0] = -1;
 
 	post_buf[0] = 0;
 	if (!fgets(post_buf, MIN(clen+1, sizeof(post_buf)), stream))
@@ -3289,10 +3306,11 @@ do_apply_cgi(const char *url, FILE *stream)
 static void
 do_upgrade_fw_cgi(const char *url, FILE *stream)
 {
-	if (f_exists(FW_IMG_NAME)) {
+	if (f_exists(FW_IMG_NAME) && get_login_safe()) {
 		notify_rc("flash_firmware");
 		websApply(stream, "Updating.asp");
 	} else {
+		unlink(FW_IMG_NAME);
 		websApply(stream, "UpdateError.asp");
 	}
 }
@@ -3303,7 +3321,7 @@ do_restore_nv_cgi(const char *url, FILE *stream)
 	char *upload_file = PROFILE_FIFO_UPLOAD;
 	int ret = -1;
 
-	if (f_exists(upload_file)) {
+	if (f_exists(upload_file) && get_login_safe()) {
 		doSystem("killall %s %s", "-q", "watchdog");
 		sleep(1);
 		ret = eval("/usr/sbin/nvram", "restore", upload_file);
@@ -3312,8 +3330,9 @@ do_restore_nv_cgi(const char *url, FILE *stream)
 			eval("/sbin/watchdog");
 		} else
 			nvram_commit();
-		unlink(upload_file);
 	}
+
+	unlink(upload_file);
 
 	/* Reboot if successful */
 	if (ret == 0) {
@@ -3330,10 +3349,11 @@ do_restore_st_cgi(const char *url, FILE *stream)
 	const char *upload_file = STORAGE_FIFO_FILENAME;
 	int ret = -1;
 
-	if (f_exists(upload_file)) {
+	if (f_exists(upload_file) && get_login_safe()) {
 		ret = eval("/sbin/mtd_storage.sh", "restore");
-		unlink(upload_file);
 	}
+
+	unlink(upload_file);
 
 	if (ret == 0) {
 		websApply(stream, "UploadDone.asp");
